Add page router
