+++ package.json
@@ -16,7 +16,9 @@
     "moment": "^2.29.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-google-forms-hooks": "^1.2.4"
+    "react-google-forms-hooks": "^1.2.4",
+    "react-icons": "^4.10.1",
+    "react-router-dom": "^6.14.1"
   },
   "devDependencies": {
     "@types/react": "^18.2.14",
