+++ src/components/FormField.tsx
@@ -1,6 +1,22 @@
 import ShortAnswerInput from './ShortAnswerInput';
 import DateInput from './DateInput';
-import { FormField } from '../types/types';
+
+type FormField = (
+  | {
+      label: string;
+      description: null;
+      type: string;
+      id: string;
+      required: boolean;
+    }
+  | {
+      label: string;
+      description: null;
+      type?: undefined;
+      id?: undefined;
+      required?: undefined;
+    }
+)[];
 
 type FormFieldProps = {
   fields: FormField;
