+++ src/pages/router.types.ts
@@ -1,0 +1,5 @@
+export interface RouterType {
+  title: string;
+  path: string;
+  element: JSX.Element;
+}
