--- src/types/types.ts
@@ -1,16 +1,0 @@
-export type FormField = (
-  | {
-      label: string;
-      description: null;
-      type: string;
-      id: string;
-      required: boolean;
-    }
-  | {
-      label: string;
-      description: null;
-      type?: undefined;
-      id?: undefined;
-      required?: undefined;
-    }
-)[];
